Created Notes.org for keeping valuable info
This entry is the body and will only be 70 characters in terms of
wrapping. This is set with the variable fill-column. Thus we can
adjust on future projects if the standards require.

We needed a notes file to keep track of the few bits of info we wanted
to retain specifically.

If there is an  open issue in the issue tracker this relates to you
add at the bottom

Resolves: #1
also #342  it would seem that magit might create links to issues. It's
a dumb link that assumes the issue exists it would seem but still very cool.